changed int to long in receive file

+++ FileHandler.cs
@@ -22,7 +22,7 @@ namespace FileSync
                     {
                         byte[] buffer = new byte[8192];
                         int read;
-                        int bytesSoFar = 0; //Use this to keep track of how many bytes have been read
+                        long bytesSoFar = 0; //Use this to keep track of how many bytes have been read
                         
 
                         do
